+++ commons/tornium_commons/models/notification_trigger.py
@@ -18,7 +18,6 @@ from peewee import (
     CharField,
     DateTimeField,
     ForeignKeyField,
-    IntegerField,
     TextField,
     UUIDField,
 )