+++ commons/tornium_commons/models/organized_crime_slot.py
@@ -14,12 +14,9 @@
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 from peewee import (
-    BigIntegerField,
     BooleanField,
     CharField,
-    DateTimeField,
     ForeignKeyField,
-    IntegerField,
     SmallIntegerField,
     UUIDField,
 )