+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added support for advanced search to `tornium-estimate.user.js`
 - Added new Lua+Solid based notification infrasture
 - Added Oban and ObanWeb to Elixir worker
+- Added OC delay notifications
+- Added OC missing tool notifications
 
 ### Changed
 - Re-enabled Discord-based authentication